refactor: rename interfaces to _serialization
